resumen y users: añadir validaciones al formulario de actualizacion

+++ database/migrations/2026_03_09_111103_create_resumens_table.php
@@ -38,7 +38,7 @@ return new class extends Migration
             //detalles tierra
             $table->integer('temp');
             $table->integer('API');
-            $table->integer('agua_sedimento');
+            $table->decimal('agua_sedimento', 6, 2);
             $table->integer('azufre');
 
             //datos buque
